requirements.txt
discord.py>=2.6.3
asyncio>=4.0.0
nest_asyncio>=1.6.0
PyNaCl>=1.6.0
asyncpg>=0.30.0